refactor: 네이밍 수정 및 함수 리팩터링
