+++ app/src/main/java/woowacourse/paint/MainViewModel.kt
@@ -3,6 +3,8 @@ package woowacourse.paint
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import woowacourse.paint.canvas.CustomColor
+import woowacourse.paint.model.ColorUiModel
 
 class MainViewModel : ViewModel() {
     private val _changingState = MutableLiveData<ChangingState>(ChangingState.Nothing)
@@ -10,11 +12,14 @@ class MainViewModel : ViewModel() {
         get() = _changingState
 
     private val _colors =
-        MutableLiveData(CustomColor.getAllColors().map { ColorUiModel(it, false) })
+        MutableLiveData(
+            CustomColor.getAllColors()
+                .map { if (it.ordinal == 0) ColorUiModel(it, true) else ColorUiModel(it, false) },
+        )
     val colors: LiveData<List<ColorUiModel>>
         get() = _colors
 
-    fun setChangeColor() {
+    fun showColorSetting() {
         if (_changingState.value == ChangingState.ColorChanging) {
             _changingState.value = ChangingState.Nothing
             return
@@ -22,7 +27,7 @@ class MainViewModel : ViewModel() {
         _changingState.value = ChangingState.ColorChanging
     }
 
-    fun setChangeThickness() {
+    fun showWidthSetting() {
         if (_changingState.value == ChangingState.WidthChanging) {
             _changingState.value = ChangingState.Nothing
             return
@@ -30,17 +35,17 @@ class MainViewModel : ViewModel() {
         _changingState.value = ChangingState.WidthChanging
     }
 
-    fun pickColor(color: ColorUiModel) {
+    fun pickColor(model: ColorUiModel) {
         _colors.value?.let { colors ->
-            val newPalette = colors.toMutableList()
-            newPalette.replaceAll {
-                if (it.color == color.color) {
+            val newColors = colors.toMutableList()
+            newColors.replaceAll {
+                if (it.color == model.color) {
                     it.copy(isPicked = true)
                 } else {
                     it.copy(isPicked = false)
                 }
             }
-            _colors.value = newPalette
+            _colors.value = newColors
         }
     }
 }
