+++ app/src/main/java/woowacourse/paint/canvas/CustomColor.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint
+package woowacourse.paint.canvas
 
 import android.graphics.Color
 
@@ -6,6 +6,6 @@ enum class CustomColor(val colorCode: Int) {
     RED(Color.RED), ORANGE(Color.parseColor("#FF9802")), YELLOW(Color.YELLOW), GREEN(Color.GREEN), BLUE(Color.BLUE);
 
     companion object {
-        fun getAllColors() = listOf(RED, ORANGE, YELLOW, GREEN, BLUE)
+        fun getAllColors() = CustomColor.values().toList()
     }
 }
