+++ app/src/main/res/layout/activity_main.xml
@@ -18,8 +18,8 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <woowacourse.paint.PaintView
-            android:id="@+id/pv_paper"
+        <woowacourse.paint.canvas.CanvasView
+            android:id="@+id/cv_canvas"
             android:layout_width="0dp"
             android:layout_height="0dp"
             app:layout_constraintBottom_toTopOf="@id/rs_thickness_changer"
@@ -31,7 +31,7 @@
             android:id="@+id/rs_thickness_changer"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:visibility="@{viewModel.changingState instanceof ChangingState.ThicknessChanging ? View.VISIBLE : View.GONE}"
+            android:visibility="@{viewModel.changingState instanceof ChangingState.WidthChanging ? View.VISIBLE : View.GONE}"
             app:layout_constraintBottom_toTopOf="@id/rv_colors"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent" />
@@ -53,7 +53,7 @@
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_margin="10dp"
-            android:onClick="@{() -> viewModel.setChangeColor()}"
+            android:onClick="@{() -> viewModel.showColorSetting()}"
             android:text="@string/main_change_color"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toStartOf="@id/btn_change_thickness"
@@ -65,7 +65,7 @@
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_margin="10dp"
-            android:onClick="@{() -> viewModel.setChangeThickness()}"
+            android:onClick="@{() -> viewModel.showWidthSetting()}"
             android:text="@string/main_change_thickness"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
